feat: separate info and status page
